requirements.txt
altair==5.5.0
python-dotenv==1.0.1
altex==0.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
entrypoints==0.4
Faker==37.4.0
favicon==0.7.0
fonttools==4.58.5
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
htbuilder==0.9.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
narwhals==1.27.1
ngrok==1.4.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.2.0
prometheus_client==0.22.1
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pymdown-extensions==10.16
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sqlparse==0.5.3
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.42.2
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
validators==0.35.0
watchdog==6.0.0
